refactor(Entire-Project): Extract all entrypoints to seperate routers
