+++ src/fastapi_server.py
@@ -1,29 +1,28 @@
 import asyncio
 import io
 import os
-import pickle
-import re
-from collections import defaultdict
-from datetime import datetime
-from typing import Dict, List
+from typing import List
 
-import exifread
 import uvicorn
 from fastapi import FastAPI, HTTPException, Query, exceptions, status
-from fastapi.responses import HTMLResponse, JSONResponse, StreamingResponse
+from fastapi.responses import HTMLResponse, StreamingResponse
 from fastapi.staticfiles import StaticFiles
 from loguru import logger
 from PIL import Image
-from pydantic import BaseModel
 
-from src.routers import face_processing
+from src.routers import (
+    classify_page_entrypoints,
+    face_processing,
+    groups_page_entrypoints,
+    image_managment,
+)
 
 from .config import AppConfig
 from .services.face_reid import FaceRecognitionService
 from .services.faces_db_service import FaceDBService
-from .services.groups_db import GROUPED_FILE, load_groups_from_file
+from .services.groups_db import GROUPED_FILE
 from .services.redis_service import RedisInterface
-from .utils.model_pydantic import Face, GroupMetadata, ImageMetadata
+from .utils.model_pydantic import Face
 
 app_config = AppConfig()
 app_config = AppConfig()
@@ -36,6 +35,20 @@ BASE_PATH = "/images"
 
 redis_service = None
 face_recognition_service = None
+
+
+classify_router = classify_page_entrypoints.ClassifyRouter()
+classify_router.create_entry_points(app)
+
+groups_router = groups_page_entrypoints.GroupsRouter()
+groups_router.create_entry_points(app)
+
+image_router = image_managment.ImagesProcessing(
+    images_base_path=BASE_PATH,
+    pickle_file_path=PICKLE_FILE,
+    group_file_path=GROUPED_FILE,
+)
+image_router.create_entry_points(app)
 try:
     redis_service = RedisInterface(host=app_config.REDIS_URL)
     face_db_service = FaceDBService(db_path=FACE_DB)
@@ -55,32 +68,6 @@ try:
 except Exception as err:
     logger.error(f"Failed to initialize redis or face recognition service: {err}")
 
-
-# Define Pydantic model for paginated response
-class PaginatedGroupsResponse(BaseModel):
-    total_groups: int
-    current_page: int
-    page_size: int
-    groups: List[GroupMetadata]
-
-
-class ToggleGroupSelection(BaseModel):
-    group_name: str
-    selection: str
-
-
-class UpdateClassificationRequest(BaseModel):
-    group_name: str
-    image_name: str
-    classification: str
-
-
-class UpdateRonInImageRequest(BaseModel):
-    group_name: str
-    image_name: str
-    ron_in_image: bool
-
-
 # Serve static files
 app.mount("/static", StaticFiles(directory=STATIC_FOLDER_LOCATION), name="static")
 app.mount("/images", StaticFiles(directory=BASE_PATH), name="images")
@@ -97,339 +84,6 @@ async def get_index(success: bool = False):
     # Read and serve the HTML file
     with open(os.path.join(STATIC_FOLDER_LOCATION, "groups.html")) as f:
         return HTMLResponse(content=f.read(), status_code=200)
-
-
-def extract_image_metadata(file: str, root: str) -> ImageMetadata:
-    """
-    Extracts metadata from a single image file.
-
-    Args:
-        file (str): The name of the image file.
-        root (str): The directory path where the file is located.
-
-    Returns:
-        ImageMetadata: An object containing metadata of the image.
-    """
-    full_path = os.path.join(root, file)
-    size = os.path.getsize(full_path)
-    type_ = file.split(".")[-1].upper()
-    creation_date = "Unknown"
-    camera = "Unknown"
-
-    # Extract EXIF metadata if available
-    with open(full_path, "rb") as image_file:
-        tags = exifread.process_file(
-            image_file, stop_tag="DateTimeOriginal", details=False
-        )
-        if "EXIF DateTimeOriginal" in tags:
-            creation_date = tags["EXIF DateTimeOriginal"].values
-        if "Image Model" in tags:
-            camera = tags["Image Model"].values
-
-    # If creation_date is still unknown, use the file's last modified date
-    if creation_date == "Unknown":
-        modified_timestamp = os.path.getmtime(full_path)
-        creation_date = datetime.fromtimestamp(modified_timestamp).strftime(
-            "%Y:%m:%d %H:%M:%S"
-        )
-
-    # Check if the file name matches the WhatsApp image naming structure
-    if camera == "Unknown":
-        whatsapp_pattern = r"IMG-\d{8}-WA\d+"
-        if re.match(whatsapp_pattern, file):
-            camera = "whatsapp"
-
-    return ImageMetadata(
-        name=file,
-        full_client_path=full_path.replace(BASE_PATH, "/images"),
-        size=size,
-        type=type_,
-        camera=camera,
-        creationDate=creation_date,
-    )
-
-
-def save_groups(groups: Dict[str, List[Dict]]):
-    """
-    Save the groups to the grouped metadata file after sorting by creation date.
-
-    Args:
-        groups (Dict[str, List[Dict]]): The dictionary of groups to save.
-    """
-    # Load existing grouped metadata if it exists
-    if os.path.exists(GROUPED_FILE):
-        with open(GROUPED_FILE, "rb") as f:
-            grouped_metadata = pickle.load(f)
-    else:
-        grouped_metadata = []
-
-    # Update or add new groups to grouped metadata
-    existing_group_names = {group["group_name"]: group for group in grouped_metadata}
-    for group_key, images in groups.items():
-        # Sort images by creation date
-        images.sort(key=lambda x: x.get("creationDate", "Unknown"))
-        representative_image = images[0]  # Select the first image as representative
-        group_thumbnail_url = representative_image.get("full_client_path")
-
-        if group_key in existing_group_names:
-            # Update existing group
-            existing_group = existing_group_names[group_key]
-
-            # Preserve existing group's selection field
-            selection = existing_group.get("selection")
-
-            # Update list_of_images, preserving classification and ron_in_image fields if they exist
-            existing_images = {
-                img.get("image_id"): img for img in existing_group["list_of_images"]
-            }
-            for image in images:
-                image_id = image.get("image_id")
-                if image_id in existing_images:
-                    existing_image = existing_images[image_id]
-                    image["classification"] = existing_image.get(
-                        "classification", image.get("classification")
-                    )
-                    image["ron_in_image"] = existing_image.get(
-                        "ron_in_image", image.get("ron_in_image")
-                    )
-
-            existing_group["list_of_images"] = images
-            existing_group["selection"] = selection
-        else:
-            # Add new group
-            group_metadata = GroupMetadata(
-                group_name=group_key,
-                group_thumbnail_url=group_thumbnail_url,
-                list_of_images=images,
-            )
-            grouped_metadata.append(group_metadata.model_dump())
-
-    # Save updated grouped metadata to a pickle file
-    sort_and_save_groups(grouped_metadata)
-
-
-def sort_and_save_groups(grouped_metadata: List[Dict]):
-    """
-    Sort the groups by date and save them to the grouped metadata file.
-
-    Args:
-        grouped_metadata (List[Dict]): The list of grouped metadata to save.
-    """
-    grouped_metadata.sort(key=lambda x: x.get("group_name", "Unknown"))
-    with open(GROUPED_FILE, "wb") as f:
-        pickle.dump(grouped_metadata, f)
-
-
-# Update the load_images function to use the new extract_image_metadata function
-@app.get("/load_images", tags=["Admin"])
-async def load_images():
-    images: List[ImageMetadata] = []
-    for root, _, files in os.walk(BASE_PATH):
-        for file in files:
-            if file.lower().endswith(
-                (".png", ".jpg", ".jpeg", ".tiff", ".bmp", ".gif")
-            ):
-                image_metadata = extract_image_metadata(file, root)
-                images.append(image_metadata)
-
-    # Convert Pydantic objects to a list of dictionaries for pickling
-    images_data = [image.model_dump() for image in images]
-    # Save the metadata to a pickle file
-    with open(PICKLE_FILE, "wb") as f:
-        pickle.dump(images_data, f)
-
-    print("Loaded images and saved metadata to pickle file.")
-
-    # Group images by date (e.g., per day)
-    groups = defaultdict(list)
-    for image in images_data:
-        # Assuming 'creationDate' is in format 'YYYY:MM:DD HH:MM:SS'
-        try:
-            date_str = image.get("creationDate", "").split(" ")[
-                0
-            ]  # Extract the date part only
-            date_obj = datetime.strptime(date_str, "%Y:%m:%d")
-            group_key = date_obj.strftime("%Y-%m-%d")  # Group by date
-        except ValueError:
-            group_key = "Unknown"
-        groups[group_key].append(image)
-
-    # Save groups using the new function
-    save_groups(groups)
-
-    return JSONResponse(
-        content={"message": "Images loaded and grouped successfully"}, status_code=200
-    )
-
-
-# Endpoint to get grouped images for preview with pagination and filtering
-@app.get(
-    "/get_groups_paginated", response_model=PaginatedGroupsResponse, tags=["Groups"]
-)
-async def get_groups_paginated(
-    page: int = Query(1, ge=1),
-    page_size: int = Query(10, ge=1),
-    filter_selections: List[str] = Query(["unprocessed"]),
-    start_date: str = Query(None),
-    end_date: str = Query(None),
-):
-    # Load grouped metadata from pickle file
-    grouped_metadata = load_groups_from_file()
-
-    # Filter groups by selection
-    grouped_metadata = [
-        group for group in grouped_metadata if group["selection"] in filter_selections
-    ]
-
-    # Filter groups by date range if specified
-    if start_date:
-        try:
-            start_date_obj = datetime.strptime(start_date, "%Y-%m-%d")
-            temp_group_metadata = []
-            for group in grouped_metadata:
-                if (
-                    group["group_name"] != "Unknown"
-                    and datetime.strptime(group["group_name"], "%Y-%m-%d")
-                    >= start_date_obj
-                ):
-                    temp_group_metadata.append(group)
-            grouped_metadata = temp_group_metadata
-        except ValueError:
-            return JSONResponse(
-                content={"error": "Invalid start_date format. Use YYYY-MM-DD."},
-                status_code=400,
-            )
-
-    if end_date:
-        try:
-            end_date_obj = datetime.strptime(end_date, "%Y-%m-%d")
-            grouped_metadata = [
-                group
-                for group in grouped_metadata
-                if datetime.strptime(group["group_name"], "%Y-%m-%d") <= end_date_obj
-            ]
-        except ValueError:
-            return JSONResponse(
-                content={"error": "Invalid end_date format. Use YYYY-MM-DD."},
-                status_code=400,
-            )
-
-    # Implement pagination
-    total_groups = len(grouped_metadata)
-    start_index = (page - 1) * page_size
-    end_index = start_index + page_size
-    paginated_groups = grouped_metadata[start_index:end_index]
-
-    # Prepare response
-    response_content = PaginatedGroupsResponse(
-        total_groups=total_groups,
-        current_page=page,
-        page_size=page_size,
-        groups=[GroupMetadata(**group) for group in paginated_groups],
-    )
-
-    return response_content
-
-
-# Endpoint to toggle group selection
-@app.post("/toggle_group_selection", tags=["Groups"])
-async def toggle_group_selection(group_select: ToggleGroupSelection):
-    # Load grouped metadata from pickle file
-    grouped_metadata = load_groups_from_file()
-
-    # Update the selection for the specified group
-    group_found = False
-    for group in grouped_metadata:
-        if group["group_name"] == group_select.group_name:
-            group["selection"] = group_select.selection
-            group_found = True
-            break
-
-    if not group_found:
-        return JSONResponse(content={"error": "Group not found"}, status_code=404)
-
-    # Save updated grouped metadata to a pickle file
-    sort_and_save_groups(grouped_metadata)
-
-    return JSONResponse(
-        content={"message": "Group selection updated successfully"}, status_code=200
-    )
-
-
-# Endpoint to update the image classification
-@app.post("/update_image_classification", tags=["Images"])
-async def update_image_classification(request: UpdateClassificationRequest):
-    grouped_metadata = load_groups_from_file()
-    # Find the group and update the classification for the specific image
-    group_found = False
-    for group in grouped_metadata:
-        if group["group_name"] == request.group_name:
-            group_found = True
-            for image in group["list_of_images"]:
-                if image["name"] == request.image_name:
-                    image["classification"] = request.classification
-                    break
-            break
-
-    if not group_found:
-        raise HTTPException(status_code=404, detail="Group not found")
-
-    # Save updated metadata
-    sort_and_save_groups(grouped_metadata)
-
-    return {"message": "Image classification updated successfully"}
-
-
-# Endpoint to update the 'Ron in the image' flag
-@app.post("/update_ron_in_image", tags=["Images"])
-async def update_ron_in_image(request: UpdateRonInImageRequest):
-    # Load existing grouped metadata
-    grouped_metadata = load_groups_from_file()
-
-    # Find the group and update the 'Ron in image' flag for the specific image
-    group_found = False
-    for group in grouped_metadata:
-        if group["group_name"] == request.group_name:
-            group_found = True
-            for image in group["list_of_images"]:
-                if image["name"] == request.image_name:
-                    image["ron_in_image"] = request.ron_in_image
-                    break
-            break
-
-    if not group_found:
-        raise HTTPException(status_code=404, detail="Group not found")
-
-    # Save updated metadata
-    sort_and_save_groups(grouped_metadata)
-
-    return {"message": "Ron in image flag updated successfully"}
-
-
-# Endpoint to get minimum and maximum dates in the groups
-@app.get("/get_min_max_dates", tags=["Groups"])
-async def get_min_max_dates():
-    grouped_metadata = load_groups_from_file()
-
-    dates = []
-    for group in grouped_metadata:
-        try:
-            date_obj = datetime.strptime(group["group_name"], "%Y-%m-%d")
-            dates.append(date_obj)
-        except ValueError:
-            continue
-
-    if not dates:
-        return JSONResponse(
-            content={"error": "No valid dates found in groups"}, status_code=404
-        )
-
-    min_date = min(dates).strftime("%Y-%m-%d")
-    max_date = max(dates).strftime("%Y-%m-%d")
-
-    return JSONResponse(
-        content={"min_date": min_date, "max_date": max_date}, status_code=200
-    )
 
 
 @app.get("/face/{face_id}/embedding", tags=["Face Recognition"])
